Reduce size

+++ ButtonHandler.ino
@@ -5,7 +5,8 @@ ButtonType needButtonRelease = ButtonType::None;
 
 // ? we can create a class or a struct instead? This can be much easier
 
-const short int buttons[3] = {ButtonType::Solid, ButtonType::Gradient, ButtonType::Dimmer};
+// ! TODO: pull into PROGMEM
+const byte buttons[3] = {ButtonType::Solid, ButtonType::Gradient, ButtonType::Dimmer};
 
 void buttonHandlerSetup()
 {
